// Progetto/src/app/old/core/mycourses/model/MyCourseDetails.ts
export interface MyCourseDetail{
    
    id_corso?: number;
    titolo?: string;
    numPosti?: number;
    numPostiDisponibili?: number
    previstoEsame?: string
    prezzo?: number
    inizio?: Date
    fine?: Date
    crediti?: number
    ore?: number
    descrizione?: string
    lingua?: string
    categoria?: string
   
 }
 